+++ ui/appui-react/src/appui-react/widget-panels/ToolSettings.tsx
@@ -6,8 +6,9 @@
  * @module ToolSettings
  */
 
-import { IModelApp } from "@itwin/core-frontend";
 import * as React from "react";
+import { IModelApp } from "@itwin/core-frontend";
+import { Text } from "@itwin/itwinui-react";
 import { UiFramework } from "../UiFramework.js";
 import { InternalFrontstageManager } from "../frontstage/InternalFrontstageManager.js";
 import { useLayout } from "../layout/base/LayoutStore.js";
@@ -17,9 +18,10 @@ import { ScrollableWidgetContent } from "../layout/widget/Content.js";
 import "./ToolSettings.scss";
 import { useActiveToolId } from "../hooks/useActiveToolId.js";
 import { useTranslation } from "../hooks/useTranslation.js";
-import { Text } from "@itwin/itwinui-react";
 import { DockedBar } from "./DockedBar.js";
 import { useActiveFrontstageDef } from "../frontstage/FrontstageDef.js";
+import { LockProvider } from "../editors/LockProvider.js";
+import { ToolSettingsEditorsProvider } from "../preview/tool-settings-lock-button/useToolSettingsLockButton.js";
 
 /** Defines a ToolSettings property entry.
  * @public
@@ -87,20 +89,24 @@ export function ToolSettingsDockedContent() {
   // for the overflow to work properly each setting in the DockedToolSettings should be wrapped by a DockedToolSetting component
   return (
     <DockedBar placement="top">
-      <DockedToolSettings
-        itemId={
-          InternalFrontstageManager.activeToolSettingsProvider?.uniqueId ??
-          "none"
-        }
-        key={forceRefreshKey}
-      >
-        {entries.map((entry, index) => (
-          <DockedToolSetting key={index}>
-            {entry.labelNode}
-            {entry.editorNode}
-          </DockedToolSetting>
-        ))}
-      </DockedToolSettings>
+      <ToolSettingsEditorsProvider>
+        <DockedToolSettings
+          itemId={
+            InternalFrontstageManager.activeToolSettingsProvider?.uniqueId ??
+            "none"
+          }
+          key={forceRefreshKey}
+        >
+          {entries.map((entry, index) => (
+            <DockedToolSetting key={index}>
+              <LockProvider>
+                {entry.labelNode}
+                {entry.editorNode}
+              </LockProvider>
+            </DockedToolSetting>
+          ))}
+        </DockedToolSettings>
+      </ToolSettingsEditorsProvider>
     </DockedBar>
   );
 }
@@ -189,9 +195,11 @@ export function ToolSettingsWidgetContent() {
       key={forceRefreshKey}
     >
       <ScrollableWidgetContent>
-        {node ?? frontstageDef?.activeToolEmptyNode ?? (
-          <EmptyToolSettingsLabel toolId={activeToolId} />
-        )}
+        <ToolSettingsEditorsProvider>
+          {node ?? frontstageDef?.activeToolEmptyNode ?? (
+            <EmptyToolSettingsLabel toolId={activeToolId} />
+          )}
+        </ToolSettingsEditorsProvider>
       </ScrollableWidgetContent>
     </div>
   );