+++ ui/components-react/src/internal.ts
@@ -2,12 +2,25 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
+export { InternalInputEditorProps } from "./components-react/editors/EditorContainer.js";
+export { registerDefaultPropertyEditor } from "./components-react/editors/PropertyEditorManager.js";
+
+export { useTranslation } from "./components-react/l10n/useTranslation.js";
+
+export {
+  CustomNumberEditor,
+  CustomNumberEditorSpec,
+} from "./components-react/new-editors/interop/old-editors/CustomNumber.js";
+
+export {
+  OldEditorMetadata,
+  isOldEditorMetadata,
+} from "./components-react/new-editors/interop/Metadata.js";
+
 export {
   addIconNodeParam,
   IconNodeEditorParams,
 } from "./components-react/properties/IconNodeEditorParam.js";
-
-export { useTranslation } from "./components-react/l10n/useTranslation.js";
 
 export { GroupTool } from "./components-react/toolbar/groupPanel/tool/Tool.js";
 export { GroupColumn } from "./components-react/toolbar/groupPanel/Column.js";