+++ ui/appui-react/CHANGELOG.md
@@ -1,5 +1,58 @@
 # Change Log - @itwin/appui-react
 
+## 5.6.0
+
+### Minor Changes
+
+- 1d29f92: Added `toolSettingsLockButton` preview feature, which when enabled will render an icon button instead of a checkbox as the default tool settings lock editor.
+
+  ```tsx
+  <PreviewFeaturesProvider
+    features={{
+      toolSettingsLockButton: true,
+    }}
+  >
+    <App />
+  </PreviewFeaturesProvider>
+  ```
+
+  Displayed icon button can be customized via the `BaseDialogItem` interface.
+
+  ```ts
+  const propertyDescription =
+    PropertyDescriptionHelper.buildLockPropertyDescription("useLength");
+  propertyDescription.displayLabel = "Toggle length lock";
+
+  const property: BaseDialogItem = {
+    value: {
+      value: true,
+    },
+    property: propertyDescription,
+    isDisabled: true,
+  };
+  // Displays an active, disabled icon button with a custom label.
+  ```
+
+  Custom lock editor components can still be displayed by customizing the lock `PropertyDescription`, even if `toolSettingsLockButton` is enabled.
+
+  ```ts
+  const property =
+    PropertyDescriptionHelper.buildLockPropertyDescription("useRadius");
+  property.editor = {
+    name: StandardEditorNames.Toggle,
+  };
+  // Displays a switch component.
+  ```
+
+- 6c97fa4: Updated the tool settings data provider to invoke `IModelApp.toolAdmin.simulateMotionEvent` after applying the tool setting property change. This causes the update to the dynamic graphics of the active tool without requiring an explicit motion event. `simulateMotionEvent` is only available from `v5.0` of `@itwin/core-frontend`.
+
+### Patch Changes
+
+- Updated dependencies [a4b74f9]
+  - @itwin/components-react@5.6.0
+  - @itwin/imodel-components-react@5.6.0
+  - @itwin/core-react@5.6.0
+
 ## 5.5.0
 
 ### Minor Changes